separate more component
