+++ tanoshi-web/src/app/component/mod.rs
@@ -19,6 +19,18 @@ pub use self::spinner::Spinner;
 pub mod model;
 
 mod filter;
+mod pager;
+pub use pager::Pager;
+
+mod reader_toolbar;
+pub use reader_toolbar::ReaderToolbar;
+
+mod reader_seekbar;
+pub use reader_seekbar::ReaderSeekbar;
+
+mod webtoon;
+pub use webtoon::Webtoon;
+
 pub use filter::Filter;
 
 use std::cell::RefCell;
