+++ tanoshi-web/src/app/mod.rs
@@ -3,7 +3,7 @@ pub use app::{App, AppRoute};
 
 pub mod browse;
 pub mod catalogue;
-pub mod chapter;
+pub mod reader;
 pub mod component;
 pub mod detail;
 pub mod history;
